# openapi/protocol-v1.yaml
openapi: 3.1.0
info:
  title: unsea Protocol API v1
  version: 1.0.0-draft
  description: |
    新一代 WhatsApp 协议层接口（基于 Baileys 7.x）。
    本规范覆盖《Baileys 协议层接口封装与 2000 账号承载方案》§ 3 + 附录 A。

    ## 设计原则

    - 协议层只负责 ws / Noise / creds / 重连，业务编排在 Java 业务层
    - 所有状态查询都带 evidence（时间戳 + 来源）
    - 所有重连相关都对应明确的语义码（PROXY_FAILED / RATE_LIMITED / NEED_REAUTH ...）
    - 接口契约稳定，protobuf 内部码不外泄

    ---

    ## 业务场景导航

    ### 🔑 账号登录 / 授权（3 种方式 + 4 种导入）

    | 业务场景 | 接口 | 说明 |
    |---|---|---|
    | 首次绑定新号（pairing code） | `POST /v1/auth/pairing-code` | 用户在手机 WA "已登录设备" 里输入 8 位码；90s 不绑会自动释放 |
    | 首次绑定新号（扫码） | `POST /v1/auth/qrcode` | 用户用手机 WA 扫返回的二维码；180s 不绑会自动释放 |
    | 标准 Baileys JSON 导入（推荐） | `POST /v1/accounts/import/baileys-json` | 完整 creds + keys，导入即可上线 |
    | 全参数登录（30+ 字段） | `POST /v1/accounts/import/params` | 移植自 Apifox `paramsLogin` |
    | 六段参数登录 | `POST /v1/accounts/import/six` | 移植自 Apifox `sixLogin`，含 deviceIdentityKey |
    | 旧 JSON 登录 | `POST /v1/accounts/import/legacy-json` | 移植自 Apifox `jsonLogin`，base64 |
    | 批量导入（最多 500） | `POST /v1/accounts/import/batch` | 含冷启动节流，避免一次性灌爆 worker |
    | 短信注册新号 | （不支持） | WA 业务不开放此能力 |

    **关键概念**：换 IP 不需要重新授权（详见 § 4.3）。一次 pairing 后 creds 永久持有，重连用本地 creds 即可。

    ### 🚪 账号生命周期

    | 业务场景 | 接口 |
    |---|---|
    | 已绑账号上线（用已存 creds） | `POST /v1/accounts/{accountId}/online` |
    | 主动下线（保留 creds，可重新 online） | `POST /v1/accounts/{accountId}/offline` |
    | 退出并移除设备 | `POST /v1/accounts/{accountId}/logout` |

    ### 🏷️ 账号状态查询

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/accounts/resolve/{accountId}:
    get:
      tags: [owner]
      summary: 查询账号当前 owner worker
      description: |
        业务层先调用此接口拿到 `ownerEndpoint`，随后直接请求 owner worker。
        协议层不做 worker 间请求代理，避免大请求/媒体请求多走一跳。
      operationId: resolveAccountOwner
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: owner 解析结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OwnerResolution' }
  /v1/accounts/resolve:
    post:
      tags: [owner]
      summary: 批量查询账号 owner worker
      description: 业务层批量调度前按 ownerEndpoint 分组，减少协议层网关压力。
      operationId: resolveAccountOwners
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountIds]
              properties:
                accountIds:
                  type: array
                  minItems: 1
                  maxItems: 1000
                  items: { type: string }
      responses:
        '200':
          description: 批量 owner 解析结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  resolvedAt: { type: string, format: date-time }
                  currentWorkerId: { type: string }
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/OwnerResolution' }

  # ──────────────────── A.3 账号状态查询 ────────────────────
...
  /v1/accounts/{accountId}/restriction:
    get:
      tags: [restriction]
      summary: 查 reachoutTimelock 状态
      description: Baileys `sock.fetchAccountReachoutTimelock()`
      operationId: getRestriction
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: 风控状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RestrictionState' }
  /v1/accounts/{accountId}/message-cap:
    get:
      tags: [restriction]
      summary: 查新 chat 配额（拉群配额）
      description: Baileys `sock.fetchNewChatMessageCap()`
      operationId: getMessageCap
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: 配额状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageCapState' }
  /v1/accounts/{accountId}/usability:
    get:
      tags: [restriction]
      summary: 业务可用性综合判定（业务侧下发任务前调一个就够）
      operationId: getUsability
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: 业务可用性
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsabilityState' }

  # ──────────────────── A.11 代理 ────────────────────
  /v1/accounts/{accountId}/proxy/bind:
    post:
      tags: [proxy]
      summary: 绑定 proxy session（首次或重新分配）
      operationId: bindProxy
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProxyBinding' }
      responses:
        '200':
          description: 绑定成功
  /v1/accounts/{accountId}/proxy/rebind:
    post:
      tags: [proxy]
      summary: 重建 socket，agent 新构造（PROXY_FAILED 处理）
      operationId: rebindProxy
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProxyBinding' }
      responses:
        '200':
          description: 已重建
  /v1/accounts/{accountId}/proxy:
    get:
      tags: [proxy]
      summary: 当前代理绑定信息
      operationId: getProxy
      parameters:
        - $ref: '#/components/parameters/AccountIdPath'
      responses:
        '200':
          description: 代理绑定信息
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProxyBindingResult' }

  # ──────────────────── A.5 群组（P0 部分完整） ────────────────────
...
  /v1/profile/{jid}/picture-url:
    get: { tags: [profile], summary: "[P2] 获取他人头像 URL", operationId: profilePictureUrl, parameters: [{ in: path, name: jid, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }

  # ──────────────────── A.9 Business 资料（P3 完整 schema） ────────────────────
...
components:
  parameters:
    AccountIdPath:
      in: path
      name: accountId
      required: true
      schema: { type: string }
      example: "acc_001"
...
  schemas:
...
    Evidence:
      type: object
      description: 状态来源证据（详见 § 4.7.1 三层证据模型）
      properties:
        wsOpen: { type: boolean }
        connectionField:
          type: string
          enum: [connecting, open, close]
        lastDateRecv: { type: string, format: date-time }
        lastPingAckAt: { type: string, format: date-time }
        ageMs: { type: integer }
        keepAliveIntervalMs: { type: integer, example: 30000 }
        slotReleased:
          type: boolean
          description: true 表示该账号当前不占 worker 在线容量
        reason:
          type: string
          nullable: true
          description: slotReleased 或状态变更原因
        updatedAt:
          type: string
          format: date-time
          description: runtime 状态更新时间
    RoutingInfo:
      type: object
      required: [ownerWorkerId, currentWorkerId, local]
      properties:
        ownerWorkerId:
          type: string
          description: accountId 当前归属 worker
        ownerEndpoint:
          type: string
          nullable: true
          description: 业务层可直连的 worker 地址；由 worker 的 PUBLIC_ENDPOINT 注册
        currentWorkerId:
          type: string
          description: 当前响应请求的 worker
        local:
          type: boolean
          description: true 表示当前 worker 就是 owner
    OwnerResolution:
      allOf:
        - $ref: '#/components/schemas/RoutingInfo'
        - type: object
          required: [accountId, assigned, resolvedAt]
          properties:
            accountId: { type: string }
            assigned:
              type: boolean
              description: false 表示账号尚未分配 owner
            resolvedAt: { type: string, format: date-time }

    # ──── 代理 ────
    ProxyBinding:
      type: object
      required: [protocol, url, sessionId, country]
      properties:
        protocol:
          type: string
          enum: [socks5, http]
        url:
          type: string
          description: 完整代理 URL（含 session 参数），例如 `socks5://brd-customer-XXX-zone-residential-session-acc_001-country-us:PASS@proxy.example.com:22225`
        sessionId:
          type: string
          description: 持久绑定账号一辈子的 session ID
        provider:
          type: string
          enum: [bright_data, smartproxy, iproyal, custom]
        country: { type: string, example: "US" }
        region: { type: string, nullable: true }
        asn: { type: string, nullable: true, example: "AS7922" }
        tier:
          type: string
          enum: [pairing, standard]
        stickyDurationSec: { type: integer, example: 600 }
    ProxyBindingResult:
      allOf:
        - $ref: '#/components/schemas/ProxyBinding'
        - type: object
          properties:
            boundAt: { type: string, format: date-time }
            lastRotatedAt: { type: string, format: date-time }
            status:
              type: string
              enum: [active, migrating, blacklisted, dead]
...
    RestrictionState:
      type: object
      required: [accountId, isActive, fetchedAt]
      properties:
        accountId: { type: string }
        isActive: { type: boolean }
        restrictedUntil: { type: string, format: date-time, nullable: true }
        enforcementType:
          type: string
          enum:
            - DEFAULT
            - BIZ_QUALITY
            - WEB_COMPANION_ONLY
            - BIZ_COMMERCE_VIOLATION_ALCOHOL
            - BIZ_COMMERCE_VIOLATION_ADULT
            - BIZ_COMMERCE_VIOLATION_ANIMALS
            - BIZ_COMMERCE_VIOLATION_BODY_PARTS_FLUIDS
            - BIZ_COMMERCE_VIOLATION_DATING
            - BIZ_COMMERCE_VIOLATION_DIGITAL_SERVICES_PRODUCTS
            - BIZ_COMMERCE_VIOLATION_DRUGS
            - BIZ_COMMERCE_VIOLATION_DRUGS_ONLY_OTC
            - BIZ_COMMERCE_VIOLATION_GAMBLING
            - BIZ_COMMERCE_VIOLATION_HEALTHCARE
            - BIZ_COMMERCE_VIOLATION_REAL_FAKE_CURRENCY
            - BIZ_COMMERCE_VIOLATION_SUPPLEMENTS
            - BIZ_COMMERCE_VIOLATION_TOBACCO
            - BIZ_COMMERCE_VIOLATION_VIOLENT_CONTENT
            - BIZ_COMMERCE_VIOLATION_WEAPONS
        raw:
          type: object
          additionalProperties: true
        fetchedAt: { type: string, format: date-time }
    MessageCapState:
      type: object
      required: [accountId, cappingStatus, fetchedAt]
      properties:
        accountId: { type: string }
        totalQuota: { type: integer }
        usedQuota: { type: integer }
        remaining: { type: integer }
        cycleStart: { type: string, format: date-time }
        cycleEnd: { type: string, format: date-time }
        cappingStatus:
          type: string
          enum: [NONE, FIRST_WARNING, SECOND_WARNING, CAPPED]
        oteStatus:
          type: string
          enum: [NOT_ELIGIBLE, ELIGIBLE, ACTIVE_IN_CURRENT_CYCLE, EXHAUSTED]
        mvStatus:
          type: string
          enum: [NOT_ELIGIBLE, NOT_ACTIVE, ACTIVE, ACTIVE_UPGRADE_AVAILABLE]
        fetchedAt: { type: string, format: date-time }
    UsabilityState:
      type: object
      required: [accountId, state, evidence]
      properties:
        accountId: { type: string }
        state:
          type: string
          enum: [ONLINE, STALE, OFFLINE, NEED_REAUTH, PROXY_FAILED, RATE_LIMITED]
        canSendText: { type: boolean }
        canSendNewChat: { type: boolean }
        canCreateGroup: { type: boolean }
        canAddToGroup: { type: boolean }
        canSendMedia: { type: boolean }
        canFollowChannel: { type: boolean }
        blockedReason:
          type: string
          nullable: true
          enum:
            - null
            - REACHOUT_TIMELOCK
            - NEW_CHAT_CAPPED
            - RATE_LIMITED
            - OFFLINE
            - NEED_REAUTH
        blockedUntil: { type: string, format: date-time, nullable: true }
        restriction: { $ref: '#/components/schemas/RestrictionState' }
        messageCap: { $ref: '#/components/schemas/MessageCapState' }
        evidence: { $ref: '#/components/schemas/Evidence' }

    # ──── 群 ────
...